requirements.txt
torch
torchvision
pandas
numpy
scikit-learn
Pillow
openpyxl
streamlit
